requirements.txt
python-telegram-bot==20.0
google-api-python-client==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
cryptography==41.0.5
python-dotenv==0.19.2
setuptools<81
reportlab>=3.6.0
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
websockets
psutil
pytest>=7.0.0
Telethon>=1.30.0
yt-dlp>=2024.4.9